refactor: extract text area for reuse
