+++ src/ui/annotation_schema_tab.css
@@ -244,7 +244,7 @@
   margin-bottom: 2px;
 }
 
-.neuroglancer-annotation-schema-textarea {
+.neuroglancer-annotation-textarea {
   height: auto;
   padding: 0;
   overflow-y: auto;
